Refactor modals in CreateTicket.vue and related components for improved state management and cleaner code
